+++ gradle-plugin-testing/build.gradle
@@ -5,7 +5,6 @@ apply plugin: 'com.palantir.gradle-plugin-testing'
 apply plugin: 'com.palantir.external-publish-gradle-plugin'
 
 dependencies {
-    implementation 'org.junit.jupiter:junit-jupiter-api'
     implementation project(':plugin-testing-core')
     implementation project(':discover-tests-cli')
     implementation 'com.fasterxml.jackson.core:jackson-databind'
